requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.1
numpy>=1.26
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
aiohttp>=3.9.5
